requirements.txt
numpy==1.23.5
Pillow
soundfile